renamed package according to sonarqube report
