+++ src/main/java/de/eberln/countdown/CountdownPiApplication.java
@@ -1,4 +1,4 @@
-package de.eberln.schleifenPi;
+package de.eberln.countdown;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
