+++ src/main/java/de/eberln/countdown/controller/CountdownController.java
@@ -1,4 +1,4 @@
-package de.eberln.schleifenPi.controller;
+package de.eberln.countdown.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
