+++ src/main/java/de/eberln/countdown/datahandling/SettingsRepository.java
@@ -1,4 +1,4 @@
-package de.eberln.schleifenPi.datahandling;
+package de.eberln.countdown.datahandling;
 
 import java.io.IOException;
