+++ src/main/java/de/eberln/countdown/errorhandling/ErrorResponse.java
@@ -1,4 +1,4 @@
-package de.eberln.schleifenPi.errorhandling;
+package de.eberln.countdown.errorhandling;
 
 import java.util.Date;
