+++ src/main/java/de/eberln/countdown/errorhandling/RestResponseEntityExceptionHandler.java
@@ -1,4 +1,4 @@
-package de.eberln.schleifenPi.errorhandling;
+package de.eberln.countdown.errorhandling;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
